DarkMode Resolution Somewhat Finished, Moved Scholarship to Scholarship Folder
