+++ lib/models/from_postgres/scholarship/scholarship.dart
